+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1-SNAPSHOT
 
 Enhancements:
+- [#3](https://github.com/caeos/snoozle/issues/3) Auto-create entity folders
 - [#4](https://github.com/caeos/snoozle/issues/4) Implement Jackson performance tips
 
 ## 0.1.0 (2018-12-03)